Fix AMP-induced NaN/inf loss and add gradient clipping to training
The heatmap sigmoid output runs in float16 under autocast; clamping it
to (1e-6, 1-1e-6) in fp16 rounds the upper bound to exactly 1.0, so
log(1 - pred) computes log(0) = -inf whenever a peak pixel is
confidently predicted. Cast to float32 before clamping. Also add
gradient-norm clipping so an unstable batch can no longer permanently
knock training out of a good region (observed: loss recovered to
mean_best_iou=0.37 by epoch 16, then diverged to precision/recall 0 by
epoch ~55 and never came back).

+++ CNN/event_maritime/losses.py
@@ -4,7 +4,11 @@ import torch
 
 
 def focal_loss(pred, target, alpha=2.0, beta=4.0):
-    pred = pred.clamp(1e-6, 1 - 1e-6)
+    # Cast to float32 before clamping: under AMP, pred arrives as float16,
+    # which can't represent 1 - 1e-6 distinctly from 1.0, so the clamp's
+    # upper bound silently becomes 1.0 and log(1 - pred) below produces -inf.
+    pred = pred.float().clamp(1e-6, 1 - 1e-6)
+    target = target.float()
     pos_mask = (target == 1).float()
     neg_mask = (target < 1).float()
     neg_weight = torch.pow(1 - target, beta)
